+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 
 # <br><b>Changelog</b><br>
 
+## ... `v1.6.7`
+* restructured the object returned from `Console.get_args()`:<br>
+  before, you accessed an arg's result with `args["<arg_alias>"]["value"]` and `args["<arg_alias>"]["exists"]`<br>	
+  now, you can directly access the result with `args.<arg_alias>.value` and `args.<arg_alias>.exists`<br>
+
 ## 17.02.2025 `v1.6.6`
 * added a new method `Console.multiline_input()`
 * added two new params to the method `Console.log_box()`:<br>